refactor: move forwardcommands into separate file to avoid duplicate code
