+++ packages/renderer-process/src/parts/ForwardCommand/ForwardCommand.js
@@ -1,0 +1,30 @@
+import * as ForwardViewletCommand from '../ForwardViewletCommand/ForwardViewletCommand.js'
+
+export const close = ForwardViewletCommand.forwardViewletCommand('close')
+export const closeEditor = ForwardViewletCommand.forwardViewletCommand('closeEditor')
+export const focus = ForwardViewletCommand.forwardViewletCommand('focus')
+export const focusIndex = ForwardViewletCommand.forwardViewletCommand('focusIndex')
+export const focusNext = ForwardViewletCommand.forwardViewletCommand('focusNext')
+export const focusPrevious = ForwardViewletCommand.forwardViewletCommand('focusPrevious')
+export const handleBlur = ForwardViewletCommand.forwardViewletCommand('handleBlur')
+export const handleClick = ForwardViewletCommand.forwardViewletCommand('handleClick')
+export const handleClickAt = ForwardViewletCommand.forwardViewletCommand('handleClickAt')
+export const handleClickClose = ForwardViewletCommand.forwardViewletCommand('handleClickClose')
+export const handleContextMenu = ForwardViewletCommand.forwardViewletCommand('handleContextMenu')
+export const handleDoubleClick = ForwardViewletCommand.forwardViewletCommand('handleDoubleClick')
+export const handleDragOver = ForwardViewletCommand.forwardViewletCommand('handleDragOver')
+export const handleDrop = ForwardViewletCommand.forwardViewletCommand('handleDrop')
+export const handleFocus = ForwardViewletCommand.forwardViewletCommand('handleFocus')
+export const handleInput = ForwardViewletCommand.forwardViewletCommand('handleInput')
+export const handlePointerDown = ForwardViewletCommand.forwardViewletCommand('handlePointerDown')
+export const handleResizerClick = ForwardViewletCommand.forwardViewletCommand('handleResizerClick')
+export const handleResizerMove = ForwardViewletCommand.forwardViewletCommand('handleResizerMove')
+export const handleScrollBarClick = ForwardViewletCommand.forwardViewletCommand('handleScrollBarClick')
+export const handleScrollBarThumbPointerMove = ForwardViewletCommand.forwardViewletCommand('handleScrollBarThumbPointerMove')
+export const handleTabClick = ForwardViewletCommand.forwardViewletCommand('handleTabClick')
+export const handleTabContextMenu = ForwardViewletCommand.forwardViewletCommand('handleTabContextMenu')
+export const handleTouchEnd = ForwardViewletCommand.forwardViewletCommand('handleTouchEnd')
+export const handleTouchMove = ForwardViewletCommand.forwardViewletCommand('handleTouchMove')
+export const handleTouchStart = ForwardViewletCommand.forwardViewletCommand('handleTouchStart')
+export const updateEditingValue = ForwardViewletCommand.forwardViewletCommand('updateEditingValue')
+export const handleBeforeInput = ForwardViewletCommand.forwardViewletCommand('handleBeforeInput')
