+++ packages/renderer-process/src/parts/ViewletExplorer/ViewletExplorerFunctions.js
@@ -1,10 +1,1 @@
-import * as ForwardViewletCommand from '../ForwardViewletCommand/ForwardViewletCommand.js'
-
-export const focus = ForwardViewletCommand.forwardViewletCommand('focus')
-export const handleBlur = ForwardViewletCommand.forwardViewletCommand('handleBlur')
-export const handleFocus = ForwardViewletCommand.forwardViewletCommand('handleFocus')
-export const handleDragOver = ForwardViewletCommand.forwardViewletCommand('handleDragOver')
-export const handleDrop = ForwardViewletCommand.forwardViewletCommand('handleDrop')
-export const handleClickAt = ForwardViewletCommand.forwardViewletCommand('handleClickAt')
-export const updateEditingValue = ForwardViewletCommand.forwardViewletCommand('updateEditingValue')
-export const handlePointerDown = ForwardViewletCommand.forwardViewletCommand('handlePointerDown')
+export * from '../ForwardCommand/ForwardCommand.js'
