+++ packages/renderer-process/src/parts/ViewletExtensions/ViewletExtensionsFunctions.js
@@ -1,10 +1,1 @@
-import * as ForwardViewletCommand from '../ForwardViewletCommand/ForwardViewletCommand.js'
-
-export const handleClick = ForwardViewletCommand.forwardViewletCommand('handleClick')
-export const handleFocus = ForwardViewletCommand.forwardViewletCommand('handleFocus')
-export const handleInput = ForwardViewletCommand.forwardViewletCommand('handleInput')
-export const handleScrollBarClick = ForwardViewletCommand.forwardViewletCommand('handleScrollBarClick')
-export const handleScrollBarThumbPointerMove = ForwardViewletCommand.forwardViewletCommand('handleScrollBarThumbPointerMove')
-export const handleTouchEnd = ForwardViewletCommand.forwardViewletCommand('handleTouchEnd')
-export const handleTouchMove = ForwardViewletCommand.forwardViewletCommand('handleTouchMove')
-export const handleTouchStart = ForwardViewletCommand.forwardViewletCommand('handleTouchStart')
+export * from '../ForwardCommand/ForwardCommand.js'
