+++ packages/renderer-process/src/parts/ViewletKeyBindings/ViewletKeyBindingsFunctions.js
@@ -1,7 +1,1 @@
-import * as ForwardViewletCommand from '../ForwardViewletCommand/ForwardViewletCommand.js'
-
-export const handleInput = ForwardViewletCommand.forwardViewletCommand('handleInput')
-export const handleClick = ForwardViewletCommand.forwardViewletCommand('handleClick')
-export const handleDoubleClick = ForwardViewletCommand.forwardViewletCommand('handleDoubleClick')
-export const handleResizerMove = ForwardViewletCommand.forwardViewletCommand('handleResizerMove')
-export const handleResizerClick = ForwardViewletCommand.forwardViewletCommand('handleResizerClick')
+export * from '../ForwardCommand/ForwardCommand.js'
