+++ packages/renderer-process/src/parts/ViewletMain/ViewletMainFunctions.js
@@ -1,6 +1,1 @@
-import * as ForwardViewletCommand from '../ForwardViewletCommand/ForwardViewletCommand.js'
-
-export const closeEditor = ForwardViewletCommand.forwardViewletCommand('closeEditor')
-export const handleClickClose = ForwardViewletCommand.forwardViewletCommand('handleClickClose')
-export const handleTabClick = ForwardViewletCommand.forwardViewletCommand('handleTabClick')
-export const handleTabContextMenu = ForwardViewletCommand.forwardViewletCommand('handleTabContextMenu')
+export * from '../ForwardCommand/ForwardCommand.js'
