+++ packages/renderer-process/src/parts/ViewletProblems/ViewletProblemsFunctions.js
@@ -1,5 +1,1 @@
-import * as ForwardViewletCommand from '../ForwardViewletCommand/ForwardViewletCommand.js'
-
-export const focusIndex = ForwardViewletCommand.forwardViewletCommand('focusIndex')
-export const handleClickAt = ForwardViewletCommand.forwardViewletCommand('handleClickAt')
-export const handleContextMenu = ForwardViewletCommand.forwardViewletCommand('handleContextMenu')
+export * from '../ForwardCommand/ForwardCommand.js'
