+++ packages/renderer-process/src/parts/ViewletQuickPick/ViewletQuickPickFunctions.js
@@ -1,6 +1,1 @@
-import * as ForwardViewletCommand from '../ForwardViewletCommand/ForwardViewletCommand.js'
-
-export const handleClickAt = ForwardViewletCommand.forwardViewletCommand('handleClickAt')
-export const handleInput = ForwardViewletCommand.forwardViewletCommand('handleInput')
-export const handleBlur = ForwardViewletCommand.forwardViewletCommand('handleBlur')
-export const handleBeforeInput = ForwardViewletCommand.forwardViewletCommand('handleBeforeInput')
+export * from '../ForwardCommand/ForwardCommand.js'
